% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue1.R
\name{queue1}
\alias{queue1}
\title{Use three inputs to predict response using R.}
\usage{
queue1(lambda, mu, T)
}
\arguments{
\item{lambda}{the first predictor (numeric)}

\item{mu}{the second predictor (logical)}

\item{T}{the third predictor (logical)}
}
\value{
the length Ls, the average stay time Ws and the probability of customer waiting Pwait.
}
\description{
The prediction model is described in http://www.babelgraph.org/wp/?p=358.
}
\examples{
\dontrun{
res <- queue1(4,6,1000)
}
}
